+++ src/types/index.ts
@@ -11,3 +11,4 @@ export * from "./message.js";
 export * from "./events.js";
 export * from "./status.js";
 export * from "./call.js";
+export * from "./system.js";